requirements.txt
ccxt
python-telegram-bot